refactor : change modules configuration QuillEditor
