+++ src/frontend/components/QuillEditor/QuillEditor.jsx
@@ -1,11 +1,10 @@
 import React from "react";
 import ReactQuill from "react-quill";
-import { QuillToolbar, modules, formats } from "./EditorToolbar";
+import { modules, formats } from "./ToolbarSetup";
 import "react-quill/dist/quill.snow.css";
 import "react-quill/dist/quill.core.css";
 import "./QuillEditor.css";
 const QuillEditor = ({ content, noteDataDispatch }) => {
-
   const handleChange = value => {
     noteDataDispatch({ type: "SET_CONTENT", payload: value });
   };
@@ -13,7 +12,6 @@ const QuillEditor = ({ content, noteDataDispatch }) => {
   return (
     <>
       <div className="text-editor">
-        <QuillToolbar />
         <ReactQuill
           theme="snow"
           value={content}
